Finish cleaning up RemapUtil.
